Manage Admin Ok et Desinations Display OK

+++ MVCClient/Models/Destination.cs
@@ -14,7 +14,7 @@ namespace MVCClient.Models
         public List<Ticket> TicketsSold { get; set; }
 
         public double SumSales { get; set; }
-        public double AvgSales { get; set; }
+        public int AvgSales { get; set; }
       
     }
 }
